secure_transfer_app/requirements.txt
cryptography
pycryptodome
fastapi
uvicorn[standard]
python-multipart
